Build for 1.3.1 in CI (#75)
Added DOCKER=1 for otbr bootstrap to skip installing resolvconf.

Fix the `expected submodule path 'openthread' not to be a symbolic
link` issue caused by make-firmware for nrf52840 creates a symbolic
link for openthread submodule. This is not accepted by the
git_archive_all tool.

It also adds CI for building 1.3.1 reference release.