[script] resolve sudo failure during Raspbian build in chroot (#128)
The `make-raspbian.bash` script uses QEMU and chroot to build the
image. Inside this environment, `otbr-setup.bash` script is executed
as root, which then uses `su -c "..." pi` to de-escalate to the pi
user.

On newer host systems, security hardening in QEMU and the kernel
prevents the `setuid` bit from being honored within the chroot
environment.

This causes the sudo command to fail with the following error:

sudo: effective uid is not 0, is /usr/bin/sudo on a file system with
the 'nosuid' option set or an NFS file system without root privileges?

because the pi user is unable to escalate back to root. 

Hence the user switch to pi is unnecessary. This change removes the
`su -c "..." pi` wrappers and executes scripts directly as the root
user with `sh -c`.